+++ GUI/src/Renderer/IRenderer.hpp
@@ -33,15 +33,15 @@ namespace zappy {
                 virtual void addEgg(
                     const int &eggId,
                     const int &fatherId,
-                    const size_t &x,
-                    const size_t &y
+                    const int &x,
+                    const int &y
                 ) = 0;
                 virtual void addPlayer(const game::Player &player) = 0;
 
                 virtual void updatePlayerPosition(
                     const int &id,
-                    const size_t &x,
-                    const size_t &y,
+                    const int &x,
+                    const int &y,
                     const game::Orientation &orientation
                 ) = 0;
                 virtual void updatePlayerLevel(const int &id, const size_t &level) = 0;
